Changes for tests_harness_fence in 1.0

+++ utils/test_harness/include/test_harness/test_harness_fence.hpp
@@ -17,7 +17,7 @@ ze_fence_handle_t create_fence(ze_command_queue_handle_t cmd_queue);
 void destroy_fence(ze_fence_handle_t fence);
 void reset_fence(ze_fence_handle_t fence);
 ze_result_t query_fence(ze_fence_handle_t fence);
-ze_result_t sync_fence(ze_fence_handle_t fence, uint32_t timeout);
+ze_result_t sync_fence(ze_fence_handle_t fence, uint64_t timeout);
 
 }; // namespace level_zero_tests
